Add pdg button to signed verdict overview
