+++ apps/judicial-system/web/src/routes/Judge/Confirmation/Confirmation.treat.ts
@@ -5,12 +5,6 @@ export const breakSpaces = style({
   whiteSpace: 'break-spaces',
 })
 
-export const pdfLink = style({
-  ':hover': {
-    textDecoration: 'none',
-  },
-})
-
 export const courtEndTimeContainer = style({
   marginBottom: theme.spacing[5],
   paddingBottom: theme.spacing[5],
