+++ apps/judicial-system/web/src/shared-components/PdfButton/PdfButton.treat.ts
@@ -1,0 +1,7 @@
+import { style } from 'treat'
+
+export const pdfLink = style({
+  ':hover': {
+    textDecoration: 'none',
+  },
+})
